NGSTACK-461 Refactor Redirect controller
